Adding book functionality 90% done
It adds new books to the database and all fields get filled in correctly. Still have to figure out how to add images to the image table (will do that later). If you were to pick a category Name it should correspond to the correct id and fill it in the the products table, so picking Best sellers will put the books category_ID to 2. Also changed the insert statements stock level to make testing easier (going to start on making stock level dynamic)