Book cover image implementation done
Should now be able to add images to the database and have it show for all users. Created ProductImages model and linked a relationship between that model and the products model. Added to store function to check if image has been uploaded and put into correct table. img src updated to retrieve the image linked to that books id