Remove book implementation done
Route created for delete function. delete function created in productsController. In the products controller the image from the productsImage table had to be deleted first and than the actual book. This is due to integrity violation and this was the easiest way to deal with it. Also added laravel built in function onclick="return confirm()" This pops up a simple un-customizable confirmation box. If we want to customize it, we have to manually create our own pop up confirmation box