Updated authentication and created logout functionality
First i created a function in UserController called logout, with a route::post in web to use the function. In the home page, i added a @auth statement that inputs a span (that shows welcome user) and a form with a button (for logging out). I also had to update the authentication function since logging out was causing errors. In user model i added function to retrieve password, id, and name so that laravel uses the ones we created in the database