Admin only implementation done
Only users who are admins have permission to create, edit and delete books from the database. First by creating custom middleware called AdminAccess and updating the kernel to refere to it. Then added middleware route in the web.php file and moved all the routes that only the admin should be able to use(add, edit, delete and update functionality). Also i changed User_Type in UserController to user_type. This to match the variable in the user model