Stock status now dynamically changes depending on stock level
Created ProductStatus model with a single function that defines the relationship with the products model. Updated Product model to include one-to-one relationship with productStatus model. In the ProductsController class i created a function called updateProductStatus. The stock status changes dynamically depending on the stock level. It creates the field in the productStatus table if one doesn't exist or updates when one already exists. I called this function in both the store and update functions respectively