Book added to basket table in database
Created basket table, basket model and basket Controller. Basket model connects to customer, admin and product models. Created function to add book into basket database. It checks if the user is logged in and what type of user it is. Also updates quantity dynamically, so it checks if a user has put book into the database first before it either creates one or updates it. Added alert to login page with css and script. Fixed basket view page layout slightly. Added form to show page to work on the basket functionality and updated routes for the basket also