Basket functionality 95% done
The basket page now retrieves the data from the basket table in the database. I used the showBasket and removeFromBasket functions in the basket controller to do this. Changes also made in the view as it retrieves the data and is than show. price of all quantites and total price is also shown. Have to be logged in to access basket now. I need to work on changing quantity while in the basket page (i will do that later). I moved delete button in the show page down the page as it was causing errors. Also i updated routes to get all the functions working