Basket quantity dropdown functionality
Added a updateQuantity function in the basket controller that dynamically allows the user to change the quantity in the basket with a dropdown box. This changes the basket quantity in the database too. The drop down box also only allows the user to pick based on the stock level. so if theres 9 books left, than the quantity cannot exceed 9. I made this change in the show page also