Fixed add to basket bugs
first it now checks if the stock level is equal to or less than 0 and if it is redirects with message saying 'Sorry, Item out of stock!'. Second it chekcs if the quantity exceeds the stock level, if it does it gives a message 'Unable to add more to basket!'.
Also chnaged javascript function so that red pop up message does not appear when stock level is 0