Finished the basic functionality of a recommended section
When a user clicks on a book and is directed to the 'show' page. The recommended section ("You may also be interested in..") shows books that the user may be interesed in based on tags from the book that is shown. For example, if a user clicks on a fantasy book by Chloe Pot, it will recommend books that are also either fantasy or from Chloe Pot. The filtering is in the productsController page. I have also created a basic forum page for book discussions, there is no actual functionality yet, just wanted to set it up