Wishlist Functionality done
So the wishlist functionality is done. When the user clicks on a book it adds it to the wishlist for later consideration (might make it look nicer later and add extra stuff). First created table in the database for the wishlist. Created wishlist view page, controller page and model page. The view page is very similar to the basket page, has a table that shows the books for that specific user. The controller has three methods; show, add and remove. Added button to add book into the wishlist in the show page. If user is not logged in then it will redirect back to homepage with message. I also did the css to for the wishlist aswell, for know its simiar to the style of the basket page and styling for the wishlist button in the show page. I also added a link to the wishlist to all the view pages, so that why there are 18+ changed file.