Order report functionality done
Created a page that allows the admin to see orders that have been placed in a list. Basically retrieves the data from the orders and orderDetails table and any relevant information from tables linked to those two(Such as Username, product name etc...). This included OrderDetails model page, OrderController, orderReport view page and the css for that page also