Discount Code functioanlity done
First thing i did was add Route::post to web file to apply the form in the view to the controller. I made a lot of changes to the applyDiscount method in the controller. It gives error messages if code is invalid or expired. It checks user's id for basket item and applies the discount correctly. I also changed the update quantity method so that the discount is applied again if the quantity changes. In the basket view page i added code to show the old price with a hashed through line and made the price and total price red to make it stand out more. Also added message popup in the css.