Comment Functionality 90% complete
There were many issue with allowing replying to replies hence why I took a different approach to doing this. At first i had many nested loops however it led to a limited depth of replies but utilising a partial view in the loop it applies it self leading to an unlimted depth recursively. Now need to fix the delete functionality as the button does not show now