Product Report Functionality done
Created Product Report page that shows a list of all products for the admin. The stock status changes colour depending on the stock level. If out of stock it turns red, if  low it turns orange, if in stock it turns green. This includes the table, header with links in the view page. Created basic CSS for the page. Added function called productReport in the productController to place all the necessary details in a variable called products