Order Processing Functionality done
All orders made by customers are now being processed. The admin can change the order to being processed in the order report page. Added column for in the orderReport view page called order Process and another for buttons. Added CSS that changes colour based on if the order is processed or not. Added function called updateOrderProcess in OrderController that dynamically allows the admin to change the order process. And added one line of code to the checkout function in the basketController