Forums Functionality Done
Forums functionality done took a different approach and broke down the whole functionality into smaller parts and , I have made it so that a user can input a thread name, description and author name and also a created at time which takes the time you press submit as the created at time. PLEASE NOTE: there is no CSS for the page at the present moment as it is a new functionality but i will be doing this very soon.