Previous Orders and Return Orders implementation
After several days of trial and error for this functionality and lots of debugging. This functionality is complete. Initially this was supposed to displayed on the profile modification page but it caused issues with routing as it was affecting a different functionality so i have implemented in a separate blade page. The CSS is yet to be done, I will be getting this ASAP.  Return orders also works it deletes it from the table and updates in inventory log.