Basket error/customer information error
So i added a quick if statement, so that if the basket is empty it will give a message to the user if they try to checkout out.
Also customer information was not showing their email or username. So i changed the $user variable, so that it goes through customer to find the user id information for username and email. This seemed to work